[VirtualInput] Moved and refactored MixinEntityRenderer
